Refactor into toggle and target classes
This code would be easier to understand if there was the concept of a
Target as well as a Toggle as classes. This would make it clearer when
a toggle was updating its own state or that of its target, which is
currently a bit difficult to follow.

This commit extracts the concept of a target into a class and adds
functions on it which can now be stubbed in tests for toggles and tested
independently.
